рефакторнул по чистой архетиктуре, закончил статистику по Студенту. По статистике нужно объеденить повторяющийся код. Далее сдеалть проверку по календарю на пересечение занятий (пока с предупреждением без графики)

+++ TaskFlow/Presentation/Views/Schedule/Meeting/Shared/MeetingForm.swift
@@ -32,7 +32,7 @@ struct MeetingForm: View {
     @State private var startDate: Date?
     @State private var finishDate: Date?
     @State private var completedDate: Date?
-    @State private var cost: Float
+    @State private var cost: Double
     @State private var details: String
     
     @Binding var isPresented: Bool
